% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrgram_splom.R
\name{splom_panel.pie}
\alias{splom_panel.pie}
\alias{splom_panel.shade}
\alias{splom_panel.ellipse}
\title{Panel functions for lattice corrgrams via splom}
\usage{
splom_panel.pie(
  x,
  y,
  z,
  subscripts,
  at = pretty(z),
  cor.method = "pearson",
  ...
)

splom_panel.shade(
  x,
  y,
  z,
  subscripts,
  at = pretty(z),
  col.regions = NULL,
  cor.method = "pearson",
  ...
)

splom_panel.ellipse(x, y, col.regions, cor.method = "pearson", ...)
}
\arguments{
\item{x}{Numeric coordinates from levelplot.}

\item{y}{Numeric coordinates from levelplot.}

\item{z}{Correlation values from levelplot.}

\item{subscripts}{Subscripts for lattice panel. (not used)}

\item{at}{Breaks for color levels.}

\item{cor.method}{Correlation method (default "pearson").}

\item{...}{Additional arguments passed to panel functions.}

\item{col.regions}{Color palette for shading (default NULL uses internal
red white blue palette).}
}
\description{
These functions provide custom panel methods for \code{lattice::splom()}.
}
\details{
\code{splom_panel.pie} Draws pie glyphs representing correlation coefficients,
omitting the diagonal.

\code{splom_panel.shade} Draws shaded rectangles with hash lines representing correlation
coefficients.

\code{splom_panel.ellipse} Draws ellipses representing correlation coefficients.
The position of the ellipse is determined by the position of the data,
and the shape of the ellipse is determined by the correlation.
}
\examples{
library(lattice)
pengvars <- c("bill_len", "bill_dep", "flipper_len", "body_mass")
splom(~penguins[ , pengvars], upper.panel=splom_panel.pie, pscales=0)
splom(~penguins[ , pengvars]|penguins$species, upper.panel=splom_panel.pie, pscales=0)
splom(~penguins[ , pengvars], upper.panel=splom_panel.shade, pscales=0)
splom(~penguins[ , pengvars]|penguins$species, upper.panel=splom_panel.shade, pscales=0)
splom(~penguins[ , pengvars], upper.panel=splom_panel.ellipse, pscales=0)
splom(~penguins[ , pengvars]|penguins$species, upper.panel=splom_panel.ellipse, pscales=0)
}
