% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vote}
\alias{vote}
\title{Voting correlations}
\format{A 12x12 matrix.}
\source{
Torben Iversen and David Soskice (2006). Electoral institutions and
the politics of coalitions: Why some democracies redistribute more than
others.  \emph{American Political Science Review}, 100, 165-81.  Table A2.
}
\description{
Voting correlations
}
\details{
These are the correlations of traits, where each trait is measured for 17 
developed countries (Europe, US, Japan, Australia, New Zealand).
}
\examples{

corrgram(vote, order=TRUE)

}
\references{
Using Graphs Instead of Tables.
\url{http://tables2graphs.com/doku.php?id=03_descriptive_statistics}
}

